1. 회원테이블 ->회원탈퇴시 모임장일 경우, 모임 삭제 후 탈퇴진행하라는 메시지 출력 / 일반 회원일 경우, 모입가입,좋아요,이벤트참석 테이블에서 데이터 삭제 메소드 추가
2. 이벤트 참석 테이블 기능 완료.
1)모임에 가입하지 않은 사람에게는 모임 가입 버튼 노출
2)모임에는 가입했으나 이벤트 참석하지 않은 사람에게 이벤트 참석 버튼 노출
3)모임에 가입했고 이벤트 참석도 한 사람에게는 이벤트 참석 취소 버튼 노출

3.신고 테이블
1)신고희망하는 유저 이메일 신고버튼 클릭시 우선 신고하는 유저가 신고당하는 유저를 이전에 신고한 적이 있는지 중복 신고인지 검색->있다면 신고 이력이 있다고 노출/없다면 유저 신고 등록 폼으로 이동->폼 작성 후 유저 신고 버튼 클릭시 유저 신고 가능
2)신고당한 횟수 카운트 하는 기능 구현 완료.

+++ togetherism/src/main/java/together/controller/MemberController.java
@@ -1,6 +1,7 @@
 package together.controller;
 
 import java.io.File;
+import java.util.List;
 import java.util.StringTokenizer;
 import java.util.UUID;
 
@@ -17,7 +18,9 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
 
+import together.model.Club_Member_JoinDTO;
 import together.model.MemberDTO;
+import together.service.Event_User_AttendServiceImpl;
 import together.service.MemberServiceImpl;
 
 
@@ -39,6 +42,7 @@ public class MemberController {
 		return "togetherview/member_idcheckResult";
 	}
 	
+	
 	/* nickname 중복검사 ajax함수로 처리부분 */
 	@RequestMapping(value = "/member_nickcheck.do")
 	public String member_nickcheck(@RequestParam("memnick") String member_nickname,
@@ -49,6 +53,7 @@ public class MemberController {
 		return "togetherview/member_nickcheckResult";
 	}
 	
+	
 	/* 로그인 폼 뷰 */
 	@RequestMapping(value = "/member_login.do")
 	public String member_login() {
@@ -56,12 +61,14 @@ public class MemberController {
 		// togetherview 폴더의 member_login 뷰 페이지 실행
 	}
 	
+	
 	/* email 찾기 폼 뷰*/
 	@RequestMapping(value = "/member_emailfind.do")
 	public String member_emailFind() {
 		return "togetherview/member_emailFind";
 		// togetherview 폴더의 member_emailFind 뷰 페이지 실행
 	}
+	
 	
 	/* email 찾기 완료  */
 	@RequestMapping(value = "/member_emailfind_ok.do")
@@ -81,12 +88,14 @@ public class MemberController {
 		}
 	}
 	
+	
 	/* 비밀번호 찾기 폼 뷰 */
 	@RequestMapping(value = "/member_pwfind.do")
 	public String member_pwfind() {
 		return "togetherview/member_pwFind";
 		// togetherview 폴더의 member_pwFind 뷰 페이지 실행
 	}
+	
 	
 	/* 비밀번호 찾기 완료  */
 	@RequestMapping(value = "/member_pwfind_ok.do")
@@ -140,6 +149,7 @@ public class MemberController {
 		}
 	}
 	
+	
 	/* 회원가입 폼 */
 	@RequestMapping(value = "/member_signup.do")
 	public String member_signup() {
@@ -147,6 +157,7 @@ public class MemberController {
 		return "togetherview/member_signup";
 		// togetherview 폴더의 member_signup 뷰 페이지 실행
 	}
+	
 	
 	/* 회원가입 저장 */
 	@RequestMapping(value = "/member_signup_ok.do")
@@ -368,35 +379,60 @@ public class MemberController {
 	}
 	
 	
-	/* 회원정보 삭제 완료 */
+	/* 회원정보 삭제 */
 	@RequestMapping(value = "/member_delete_ok.do")
 	public String member_del_ok(String member_pw, String member_del_reason,
 								Model model,
 							    HttpSession session) throws Exception {
-		int result = 1;
+		int result = 0;
+		
 		String email = (String) session.getAttribute("email");
+		
+		List<Club_Member_JoinDTO> hostyn = memberService.checkHost(email);
+		
+		
+		
+		//비번 비교를 위해 정보 불러옴
 		MemberDTO old = memberService.userCheck(email);
+		
 
-		if (!old.getMember_pw().equals(member_pw)) {
-			
+		if (hostyn != null) {		// 모임장일 경우
 			model.addAttribute("result", result);
 			return "togetherview/member_deleteResult";
 			
-		} else {// 비번이 같은 경우
-			
-			MemberDTO member = new MemberDTO();
-			member.setMember_email(email);;
-			member.setMember_del_reason(member_del_reason);;
+		}else {
+			if(!old.getMember_pw().equals(member_pw)) {	// 비번 불일치
+				result = 1;
+				model.addAttribute("result", result);
+				return "togetherview/member_deleteResult";
+				
+			}else {												// 비번 일치
+				
+				MemberDTO member = new MemberDTO();
+				member.setMember_email(email);;
+				member.setMember_del_reason(member_del_reason);;
+				
+				
+				//1. 모임가입 삭제
+				memberService.deleteJoinclub(email);
+				
+				//2. 좋아요 삭제
+				memberService.deleteHeart(email);
+				
+				//3. 이벤트 참석 삭제
+				memberService.deleteEventattend(email);
+				
 
-			memberService.deleteMember(member);// 삭제 메서드 호출
-			
-			result = 2;
-			
-			model.addAttribute("result", result);
+				memberService.deleteMember(member);// 삭제 메서드 호출
+				
+				result = 2;
+				
+				model.addAttribute("result", result);
 
-			session.invalidate();	// 세션만료
+				session.invalidate();	// 세션만료
 
-			return "togetherview/member_deleteResult";
+				return "togetherview/member_deleteResult";
+			}
 		}
 	}
